demo/requirements.txt
requests
flask
PyPDF2
Pillow
